fix(certinel): signal when watches are done
When closing the certinel or filesystem watches, it was possible for
a final read on the closed error channel to return the zero value, which
we would then attempt to propagate to the object's own (closed) error
channel.

This changeset adds a done channel to track the state of the watcher and
return from goroutines on close. This also allows us to better handle
close-then-watch or multiple-close cases.