fix(fswatcher): handle k8s-style symlinks
When a Secret is mounted into a container by Kubernetes, the files
created at the "volumeMount" location are symlinks to the real files in
hidden directories. When the kubelet updates the Secret, the new file is
created in a new hidden directory, a hidden symlink is created, which is
then copied over the original symlink.

The previous implementation of fswatcher was watching the original
symlink inode. When the new symlink was copied over, fswatcher observed
a "remove" event, and then saw no other updates. This changeset changes
the behavior to watch the directory, this allows us to observe the new
symlink being copied over the original symlink, and trigger a reload of
certificates.

To preserve behavior when the files are not symlinks, fswatcher checks
for create and modification events for the original file name.

Fixes #12