feat: directly implement certinels by watchers
This is a breaking change that refactors the certinel project. This
refactor had two goals:

1. Make it easier to safely use certinel. The API exposed by 0.3.x
   required users to construct a watcher and then a certinel from that
   watcher. This caused confusion on who should be calling which Watch
   function, and generally made the main function verbose. The
   Node.js-style errback function also felt extremely out of place in
   Go.

2. Make it easier to implement certinel watchers safely. The API exposed
   by 0.3.x required implementer to return loaded certificates and
   errors over channels, so that the wrapping Certinel instance could
   handle these events. This resulted in an API that was difficult to
   implement correctly, leading to many panics and race conditions
   implementing the Close function. As a result, I've not been keen in
   writing other implementations.

This refactor modifies the fswatcher implementation to directly
implement the Certinel interface. The Start function replaces the Watch
function for starting asynchronous work, whose lifetime is clearly
scoped by a context. These changes make it easier for users to use, as
they only need to construct one instance, and easier to develop, as
channels are no longer crossing package boundaries.