chore: document the atomic writer pattern
When certinels reload certificates from disk, either via timer or being
notified by the filesystem, they reload both the certificates and the
keys from disk. If the certificate and keys do not match, the certinels
exit with an error. This can cause problems if the keys and certificates
are not updated together atomically, as the watcher may trigger after
only one has been updated.

This changeset adds a test for Kubernetes's atomic writer[1] pattern,
and documents a similar pattern that users who update key and
certificates can follow.

[1]: https://github.com/kubernetes/kubernetes/blob/v1.21.3/pkg/volume/util/atomic_writer.go#L49-L54

Fixes #14