feat: support watching as Watchman client
In certinel we provide support for watching the filesystem for changed
certificates using `fswatcher`. While this works well in most cases, it
does have some limitations, such as supporting only a limited number of
concurrent watchers, or dropping events if processing is delayed.

This changeset adds supports for using Facebook's Watchman file watching
service as a certinel provider. Watchman operates as a service,
subscribing to directories on behalf of multiple clients, and better has
better handling of overflow events and other edge cases.