feat(otel): add metrics about loaded certificates
It was previously difficult to determine if a certinel implementation
was properly detecting changes and reloading certificates. This lack of
observability has made some teams uneasy of relying on certinel watchers.

This changeset adds OpenTelemetry metrics support to the fswatcher and
ticker packages. Both packages have been configures to report
"certificate.not_before_timestamp" and "certificate.not_after_timestamp"
meters, with attributes "certificate.serial" and "certificate.path".
While OpenTelemetry defines [Semanatic Conventions for TLS][0], the
current conventions are scoped for client or server certificates, which
this package is unable to determine.

[0]: https://opentelemetry.io/docs/specs/semconv/attributes-registry/tls/

Fixes: #27